Home Menu Options UI Adjustments
Slightly decreased the height of the option boxes in order to fit the boxes in one screen . Previously, it was being cut off by the top header or the bottom nav bar due to the boxes being too long.

Changed positioning of the option images slightly so the option titles do not clash with the dark colours of the background images.